requirements.txt
PyYAML
internetarchive
